+++ opyoid/bindings/registered_binding.py
@@ -1,4 +1,4 @@
-from typing import Generic, Tuple, TYPE_CHECKING
+from typing import Generic, Tuple, TYPE_CHECKING, Optional
 
 import attr
 
@@ -7,6 +7,7 @@ from opyoid.utils import InjectedT
 from .binding import Binding
 
 if TYPE_CHECKING:
+    from .abstract_module import AbstractModule
     from .private_module import PrivateModule
 
 
@@ -15,3 +16,4 @@ class RegisteredBinding(Generic[InjectedT]):
     raw_binding: Binding[InjectedT]
     source_path: Tuple["PrivateModule", ...] = attr.Factory(tuple)
     target: FrozenTarget[InjectedT] = attr.Factory(lambda self: self.raw_binding.target, takes_self=True)
+    multi_bind_parent: Optional["AbstractModule"] = None